% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snaplocs.R
\docType{data}
\name{locs}
\alias{locs}
\title{Point locations of interest in Alaska and western Canada associated with various SNAP projects and applications.}
\format{A data frame with 3914 rows and 4 variables:
\describe{
  \item{Location}{point location name}
  \item{Group}{region group/set name, a state, privince or territory}
  \item{lon}{longitude}
  \item{lat}{latitude}
}}
\source{
For a visual demonstration of where these locations lie and their availability for use in a live application,
see \url{https://uasnap.shinyapps.io/cc4liteFinal/}
}
\usage{
locs
}
\description{
This data set contains point locations and minimal associated metadata for 3,914 communities
and similar units in Alaska and western Canada. Locations include major and minor cities, towns, villages, and some other locations such as mines.
}
\keyword{datasets}
