% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{metadata}
\alias{metadata}
\alias{get_region}
\alias{get_country}
\alias{get_coords}
\title{Get point location metadata.}
\usage{
get_region(location, region)

get_country(location, region)

get_coords(location, region, keep_cols = FALSE)
}
\arguments{
\item{location}{character, a name of a location in the \code{locs} data frame. May be a vector.}

\item{region}{character, if supplied, the \code{locs} data set is first filtered to \code{region}. May be a vector.}

\item{keep_cols}{logical, keep all columns with \code{get_coords}. Defaults to \code{FALSE}, returning only the \code{lon} and \code{lat} columns.}
}
\value{
a character string, or a data frame for \code{get_coords}.
}
\description{
These functions return metadata associate with a point location.
}
\details{
\code{get_region} returns the state, province or territory of a point location.
\code{get_country} returns the name of the country (US or Canada).
\code{get_coords} returns the longitutde and latitude of a point location as a data frame.
}
\examples{
x <- "Calgary"
get_region(x)
get_country(x)
get_coords(x)
}
