% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\docType{data}
\name{regions}
\alias{regions}
\title{Common SNAP regions}
\format{A data frame with 86 rows and 2 columns giving the region name and the set/group it belongs to.}
\usage{
regions
}
\description{
Commonly used political and ecological polygon regions at SNAP.
}
\details{
A data frame of available political and ecological regions and their respective region groups/sets
commonly used with SNAP data spatial distributions and regional summary statistics.
}
\keyword{datasets}
