% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snaplocs.R
\docType{package}
\name{snaplocs}
\alias{snaplocs}
\alias{snaplocs-package}
\title{snaplocs: Functions for working with SNAP point location data.}
\description{
The \code{snaplocs} package provides functions for working with point location data used in various applications and projects by the
Scenarios Network for Alaska and Arctic Planning at the University of Alaska Fairbanks. Point location data is included.
Locations include major and minor cities, towns, villages, and some other locations such as mines.
This package is part of the SNAPverse collection of R packages.
}
